move ReplayError to mod Replay
